externalize build_str
